+++ tests/Unit/Codegen/DirectoryFinderTest.php
@@ -4,6 +4,7 @@ namespace Spawnia\Sailor\Tests\Unit\Codegen;
 
 use Spawnia\Sailor\Codegen\DirectoryFinder;
 use Spawnia\Sailor\Tests\TestCase;
+use function uniqid;
 
 final class DirectoryFinderTest extends TestCase
 {
@@ -12,6 +13,13 @@ final class DirectoryFinderTest extends TestCase
         $finder = new DirectoryFinder(__DIR__ . '/finder');
 
         self::assertCount(3, $finder->documents());
+    }
+
+    public function testCreatesDirIfNotExists(): void
+    {
+        $finder = new DirectoryFinder(__DIR__ . '/' . uniqid('finder-'));
+
+        self::assertCount(0, $finder->documents());
     }
 
     public function testNoFiles(): void