+++ src/Convert/ScalarConverter.php
@@ -2,28 +2,17 @@
 
 namespace Spawnia\Sailor\Convert;
 
+/** Does no conversion, because custom scalars are opaque without further knowledge and bespoke implementations. */
 class ScalarConverter implements TypeConverter
 {
-    public function fromGraphQL($value): string
+    public function fromGraphQL($value)
     {
-        return $this->toString($value);
+        return $value;
     }
 
-    public function toGraphQL($value): string
+    public function toGraphQL($value)
     {
-        return $this->toString($value);
-    }
-
-    /**
-     * @param mixed $value Should be string
-     */
-    protected function toString($value): string
-    {
-        if (! is_string($value)) {
-            $notString = gettype($value);
-            throw new \InvalidArgumentException("Expected string, got {$notString}");
-        }
-
+        // @phpstan-ignore-next-line Assume the developer is passing a valid value, json_encode() will crash otherwise
         return $value;
     }
 }