+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Create searchPath dir if it does not exist
+
 ## v0.29.1
 
 ### Fixed