+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## v0.29.2
+
+### Fixed
+
+- Do not assume scalar values are `string`, use `mixed`
+
 ## v0.29.1
 
 ### Fixed